Move x-grid-data-generator package under /grid in packages
