+++ packages/grid/x-grid-data-generator/src/renderer/renderAvatar.tsx
