+++ packages/grid/x-grid-data-generator/src/renderer/renderPnl.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import clsx from 'clsx';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
-import { CellParams, getThemePaletteMode } from '../utils';
+import { CellParams, getThemePaletteMode } from '../../../_modules_/grid';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
