+++ packages/grid/x-grid-data-generator/src/renderer/renderProgress.tsx
