+++ packages/grid/x-grid-data-generator/src/renderer/renderRating.tsx
