+++ packages/grid/x-grid-data-generator/src/renderer/renderTotalPrice.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import clsx from 'clsx';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
-import { CellParams, muiStyleAlpha } from '../utils';
+import { CellParams, muiStyleAlpha } from '../../../_modules_/grid';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
