--- packages/x-grid-data-generator/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './utils';
-export * from './params';
